Increase efficiency through using state for max
The previous version ran a for loop through the
entire array every re-render. To minimize resource
usage, store the data of the max anecdote in a
state instead. Update the max anecdote state
happens only when the current anecdote's votes increments
to a value greater than the vote of the present max state,
thus replacing it.