fix: Store specific c-data in component state
Initially, all results were being displayed through a synchrounous
function. This was a problem since obtaining the data for a specific
country is done asynchronously. Thus, a promise was being returned
instead of the specific country data. This caused a render loop as
whenever React tries to render the promise, it ends up rendering
nothing; BUT when the promise finally resolves, it causes a re-render
that results to the component running the async code that returns a
promise. Ultimately, this results to a loop.

To solve this, use countryResults as a dependency and if it exists and
it's length is 1, then asynchronously request for the data of the
specific country. Store the response data from the request in a state.
Then, the data can be easily displayed without any issue.